+++ lib/flutter_yandex_games.dart
@@ -1,14 +1,13 @@
-import 'dart:js';
-import 'dart:js_util';
+import 'dart:js_interop';
 
-import 'yandex_games_js_proxy.dart';
 import 'js_utils.dart';
+import 'yandex_games_js_proxy.dart';
 
 class YandexGames {
   /// Initializes the Yandex Games sdk and gets the player object.
   /// Call this on start of your application.
   static Future<bool> init() async {
-    _yaSdk = await promiseToFuture(YaGamesJs.init());
+    _yaSdk = await yaGamesInit().toDart;
     _player = await _loadPlayer();
     _loadingApi = LoadingApi(_yaSdk);
     _gameplayApi = GameplayApi(_yaSdk);
@@ -16,8 +15,8 @@ class YandexGames {
   }
 
   static Future<Player> _loadPlayer() async {
-    return Player(await promiseToFuture(
-        _yaSdk.getPlayer(GetPlayerOptions(scopes: false))));
+    return Player(
+        await _yaSdk.getPlayer(GetPlayerOptions(scopes: false)).toDart);
   }
 
   static late YaSdk _yaSdk;
@@ -30,29 +29,38 @@ class YandexGames {
   }
 
   static LoadingApi get loadingApi => _loadingApi;
+
   static GameplayApi get gameplayApi => _gameplayApi;
 
   ///Displays Fullscreen AD
   static void showFullscreenAd(
-      {Function(bool wasShown)? onClose, Function(dynamic error)? onError}) {
+      {void Function()? onOpen,
+      void Function(bool wasShown)? onClose,
+      void Function(dynamic error)? onError}) {
     _yaSdk.adv.showFullscreenAdv(ShowFullscreenAdOptions(
         callbacks: FullscreenAdCallbacks(
-            onClose: wrapCallback<Function>(onClose),
-            onError: wrapCallback<Function>(onError))));
+      onOpen: onOpen?.toJS,
+      onClose: onClose?.toJS,
+      onError: ((JSAny? error) {
+        onError?.call(mapify(error));
+      } as void Function(JSAny?)).toJS,
+    )));
   }
 
   /// Displays Rewarded Video AD and calls callback functions on Ad events
   static void showRewardedVideoAd(
-      {Function? onOpen,
-      Function? onRewarded,
-      Function? onClose,
-      Function(dynamic error)? onError}) {
+      {void Function()? onOpen,
+      void Function()? onRewarded,
+      void Function()? onClose,
+      void Function(dynamic error)? onError}) {
     _yaSdk.adv.showRewardedVideo(ShowRewardedVideoOptions(
         callbacks: RewardedVideoCallbacks(
-      onOpen: wrapCallback<Function>(onOpen),
-      onRewarded: wrapCallback<Function>(onRewarded),
-      onClose: wrapCallback<Function>(onClose),
-      onError: wrapCallback<Function>(onError),
+      onOpen: onOpen?.toJS,
+      onRewarded: onRewarded?.toJS,
+      onClose: onClose?.toJS,
+      onError: ((JSAny? error) {
+        onError?.call(mapify(error));
+      } as void Function(JSAny?)).toJS,
     )));
   }
 
@@ -66,7 +74,7 @@ class YandexGames {
   /// More info:
   /// https://yandex.ru/dev/games/doc/dg/sdk/sdk-review.html
   static Future<CanReviewResponse> canReview() async {
-    return promiseToFuture<CanReviewResponse>(_yaSdk.feedback.canReview());
+    return _yaSdk.feedback.canReview().toDart;
   }
 
   /// Displays review dialog.
@@ -74,21 +82,19 @@ class YandexGames {
   /// Returns [feedbackSent] true if user reviewed the game.
   /// And false if user closed the review dialog.
   static Future<RequestReviewResponse> requestReview() async {
-    return promiseToFuture<RequestReviewResponse>(
-        _yaSdk.feedback.requestReview());
+    return _yaSdk.feedback.requestReview().toDart;
   }
 
   /// Opens Authentication dialog.
   static Future<bool> openAuthDialog() async {
-    await promiseToFuture(_yaSdk.auth.openAuthDialog());
+    await _yaSdk.auth.openAuthDialog().toDart;
     _player = await _loadPlayer();
     return true;
   }
 
   /// Checks if it's possible to add a shortcut.
   static Future<bool> canShowShortcutPrompt() async {
-    var response = await promiseToFuture<CanShowPromptResponse>(
-        _yaSdk.shortcut.canShowPrompt());
+    var response = await _yaSdk.shortcut.canShowPrompt().toDart;
     return response.canShow;
   }
 
@@ -99,8 +105,7 @@ class YandexGames {
   ///
   /// Will return true in case of the accepted prompt.
   static Future<bool> showShortcutPrompt() async {
-    var response =
-        await promiseToFuture<ShowPromptResponse>(_yaSdk.shortcut.showPrompt());
+    var response = await _yaSdk.shortcut.showPrompt().toDart;
     return response.outcome == "accepted";
   }
 
@@ -118,7 +123,7 @@ class Player {
   Player(this.player);
 
   //Returns player's ID
-  String getUniqueID() => player.getUniqueID();
+  String getUniqueID() => player.getUniqueID().toString();
 
   /// Sets player data.
   ///
@@ -126,7 +131,7 @@ class Player {
   /// Data will not be merged with existing data.
   /// If [flush] is true sends the data immediately.
   void setData(Map<dynamic, dynamic> data, {bool flush = false}) {
-    var newData = jsify(data);
+    var newData = data.jsify();
     player.setData(newData, flush);
   }
 
@@ -135,12 +140,8 @@ class Player {
   /// Be sure to include all data when you use it.
   /// Data will not be merged with existing data.
   Future<Map<dynamic, dynamic>> getData() async {
-    JsObject data = await promiseToFuture<JsObject>(player.getData());
+    JSObject data = await player.getData().toDart;
     var map = mapify(data);
-    //Dart objects contain a JS object in 'o' field? Have to do this for now.
-    if (map["o"] != null) {
-      return map["o"];
-    }
     return map;
   }
 
@@ -148,12 +149,12 @@ class Player {
   ///
   /// Yandex Games SDK returns mode 'lite' if player is not authorized.
   bool isAuthorized() {
-    return player.getMode() != 'lite';
+    return player.isAuthorized();
   }
 }
 
 class LoadingApi {
-  YaSdk _yaSdk;
+  final YaSdk _yaSdk;
 
   LoadingApi(this._yaSdk);
 
@@ -164,7 +165,7 @@ class LoadingApi {
 }
 
 class GameplayApi {
-  YaSdk _yaSdk;
+  final YaSdk _yaSdk;
 
   GameplayApi(this._yaSdk);